I made the code 'ExpGaussianNoise', which was originally designed such that a bar was in closed-loop with regular jumps in blocks 1 and 3, and then didn't jump in block 2 (gaussian noise would then be added to the block 2, but that comes from the python code). I have now modified it, such that the bar always jumps every 100 sec, but the stimulus changes polarity from block 1 to block 2 and from block 2 to block 3.
I made a code called 'ExpOpticFlow' where the animal is in closed loop with a vertical grating. Both the x and the y dimension of the pattern are in closed-loop and the pattern was created in a way such that the yaw position of the animal can control the x dimension, and then give you rotational optic flow, while the x position of the animal can control the y dimension, and that will give you translational optic flow.

I made a code to analyze the pooled results from experiment 13 (Experiment13_PooledAnalysis), and one to analyze experiment 14 (Experiment14_analysis).

I changed the codes 'getDataAroundJump', 'runClosedLoopAnalysis', 'runOpenLoopAnalysis' and 'getDataAroundJumpFiltered' so that they now no longer add a - to the x position, since that has been corrected with the python code itself.

I created setting to run a fourier bar stimulus with Allie's flies ('settings_clockwiseFourier', 'settings_counterclockwiseFourier')

I changed the length of the 'settings_fastClockwiseOpenLoop' and 'settings_fastCounterclockwiseOpenLoop' so that they're now both the same duration.